LPD-77031 Move to a suitable utility class
